# docs/swagger.yaml
definitions:
...
  model_base.BaseArticleGroup:
    properties:
      groupId:
        description: 组id
        type: integer
      groupName:
        description: 组名
        type: string
      libraryId:
        description: 库id
        type: integer
      status:
        description: 状态
        type: string
      userId:
        description: 用户id
        type: integer
    type: object
  model_base.BaseArticleLabel:
    properties:
      labelId:
        description: 标签id
        type: integer
      labelName:
        description: 标签名
        type: string
      status:
        description: 状态
        type: string
      userId:
        description: 用户id
        type: integer
    type: object
...
  result.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /base/baseArticleGroup/:groupId:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: groupId
        in: path
        name: groupId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取组详情
      tags:
      - Base - 文章组管理
  /base/baseArticleGroup/{page}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 开始时间
        in: query
        name: endTime
        type: string
      - description: 组id
        in: query
        name: groupId
        type: integer
      - description: 组名
        in: query
        name: groupName
        type: string
      - description: 库id
        in: query
        name: libraryId
        type: integer
      - description: 当前页（非分页接口不传）
        in: query
        name: pageNum
        type: integer
      - description: 每页条数（非分页接口不传）
        in: query
        name: pageSize
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 用户id
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取组列表
      tags:
      - Base - 文章组管理
  /base/baseArticleGroup/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: baseArticleGroup
        in: body
        name: baseArticleGroup
        required: true
        schema:
          $ref: '#/definitions/model_base.BaseArticleGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 添加组
      tags:
      - Base - 文章组管理
  /base/baseArticleGroup/deletes:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: groupIds
        in: query
        name: groupIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除组（实体删除）
      tags:
      - Base - 文章组管理
  /base/baseArticleGroup/edit:
    put:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: baseArticleGroup
        in: body
        name: baseArticleGroup
        required: true
        schema:
          $ref: '#/definitions/model_base.BaseArticleGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 修改组
      tags:
      - Base - 文章组管理
  /base/baseArticleGroup/myPage:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 开始时间
        in: query
        name: endTime
        type: string
      - description: 组id
        in: query
        name: groupId
        type: integer
      - description: 组名
        in: query
        name: groupName
        type: string
      - description: 库id
        in: query
        name: libraryId
        type: integer
      - description: 当前页（非分页接口不传）
        in: query
        name: pageNum
        type: integer
      - description: 每页条数（非分页接口不传）
        in: query
        name: pageSize
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 用户id
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取当前用户的组列表
      tags:
      - Base - 文章组管理
  /base/baseArticleGroup/removes:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: groupIds
        in: query
        name: groupIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除组（逻辑删除）
      tags:
      - Base - 文章组管理
  /base/baseArticleLabel/{labelId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: labelId
        in: path
        name: labelId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取标签详情
      tags:
      - Base - 文章标签管理
  /base/baseArticleLabel/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: baseArticleLabel
        in: body
        name: baseArticleLabel
        required: true
        schema:
          $ref: '#/definitions/model_base.BaseArticleLabel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 添加标签
      tags:
      - Base - 文章标签管理
  /base/baseArticleLabel/deletes:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: labelIds
        in: query
        name: labelIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除标签（实体删除）
      tags:
      - Base - 文章标签管理
  /base/baseArticleLabel/edit:
    put:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: baseArticleLabel
        in: body
        name: baseArticleLabel
        required: true
        schema:
          $ref: '#/definitions/model_base.BaseArticleLabel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 修改标签
      tags:
      - Base - 文章标签管理
  /base/baseArticleLabel/myPage:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 开始时间
        in: query
        name: endTime
        type: string
      - description: 标签id
        in: query
        name: labelId
        type: integer
      - description: 标签名
        in: query
        name: labelName
        type: string
      - description: 当前页（非分页接口不传）
        in: query
        name: pageNum
        type: integer
      - description: 每页条数（非分页接口不传）
        in: query
        name: pageSize
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 用户id
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取当前用户的标签列表
      tags:
      - Base - 文章标签管理
  /base/baseArticleLabel/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 开始时间
        in: query
        name: endTime
        type: string
      - description: 标签id
        in: query
        name: labelId
        type: integer
      - description: 标签名
        in: query
        name: labelName
        type: string
      - description: 当前页（非分页接口不传）
        in: query
        name: pageNum
        type: integer
      - description: 每页条数（非分页接口不传）
        in: query
        name: pageSize
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 用户id
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取标签列表
      tags:
      - Base - 文章标签管理
  /base/baseArticleLabel/removes:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: labelIds
        in: query
        name: labelIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除标签（逻辑删除）
      tags:
      - Base - 文章标签管理
...
swagger: "2.0"